Move browser view to portal stack [WIP]
Initial work to move the browser view into the portal stack as a card.
Mostly works, but the effect on Android Pixel is not desirable.  When
leaving the browser card, there is a pause with a blank brown screen.

Will require a refactor of how Journal records/launches URLs.  Currently
the name of the stack indicates orientation, but if a single card is used
then that will not be present.  This could be accomplished using the same
method that initially makes the determination.

Portal library required modifications to not close the stack when
navigating to the browser card.

App stack not included, but the stack files would need to be modified to
include a the behavior for the new card.

Portal stack included with the new card.

Object placement still needs to be worked, but going to work that under
a different branch.  Merging into this one would be easy if we decide to
continue additional work here.

Issue #236